python program to find a hash file
In this program, we open the file in binary mode. Hash functions are available in the hashlib module. We loop till the end of the file using a while loop. On reaching the end, we get empty bytes object.

In each iteration, we only read 1024 bytes (this value can be changed according to our wish) from the file and update the hashing function.